+++ src/ResourceManager/Advisor/Commands.Advisor/Cmdlets/EnableAzAdvisorRecommendation.cs
@@ -19,6 +19,7 @@ namespace Microsoft.Azure.Commands.Advisor.Cmdlets
     using System.Management.Automation;
     using Microsoft.Azure.Commands.Advisor.Cmdlets.Models;
     using Microsoft.Azure.Commands.Advisor.Cmdlets.Utilities;
+    using Microsoft.Azure.Commands.Advisor.Cmdlets.Utilities.Client;
     using Microsoft.Azure.Commands.Advisor.Properties;
     using Microsoft.Azure.Commands.Advisor.Utilities;
     using Microsoft.Azure.Management.Advisor.Models;
@@ -53,14 +54,13 @@ namespace Microsoft.Azure.Commands.Advisor.Cmdlets
         /// <summary>
         /// Gets or sets the Resource Id.
         /// </summary>
-        [Parameter(ParameterSetName = "IdParameterSet", Position = 0, ValueFromPipelineByPropertyName = true, Mandatory = true, HelpMessage = "Resource Id of the recommendation to be suppressed.")]
+        [Parameter(ParameterSetName = IdParameterSet, Position = 0, ValueFromPipelineByPropertyName = true, Mandatory = true, HelpMessage = "Resource Id of the recommendation to be suppressed.")]
         public string ResourceId { get; set; }
 
         /// <summary>
         /// Gets or sets the object passed from the PowerShell piping
         /// </summary>
         [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0, ParameterSetName = InputObjectParameterSet, HelpMessage = "The powershell object type PsAzureAdvisorResourceRecommendationBase returned by Get-AzAdvisorRecommendation call.")]
-        [ValidateNotNullOrEmpty]
         public PsAzureAdvisorResourceRecommendationBase InputObject { get; set; }
 
         /// <summary>
@@ -78,12 +78,11 @@ namespace Microsoft.Azure.Commands.Advisor.Cmdlets
         public List<PsAzureAdvisorResourceRecommendationBase> SuppressionDelete(string resourceUri, string recommendationId)
         {
             AzureOperationResponse<IPage<SuppressionContract>> suppressionList = null;
-            AzureOperationResponse<IPage<ResourceRecommendationBase>> recommendationList = null;
             IEnumerable<PsAzureAdvisorSuppressionContract> psSuppressionContractList = null;
-
             IList<PsAzureAdvisorResourceRecommendationBase> responseRecommendationList = new List<PsAzureAdvisorResourceRecommendationBase>();
             List<PsAzureAdvisorResourceRecommendationBase> responseRecommendation = new List<PsAzureAdvisorResourceRecommendationBase>();
             List<AzureOperationResponse> response = new List<AzureOperationResponse>();
+            RecommendationResource recommendationResourceUtil = new RecommendationResource();
 
             // Get the list of all suppressions
             suppressionList = this.ResourceAdvisorClient.Suppressions.ListWithHttpMessagesAsync().Result;
@@ -100,8 +99,7 @@ namespace Microsoft.Azure.Commands.Advisor.Cmdlets
             }
 
             // Get all the recommendation and convert to its corresponding psobject
-            recommendationList = this.ResourceAdvisorClient.Recommendations.ListWithHttpMessagesAsync().Result;
-            responseRecommendationList = PsAzureAdvisorResourceRecommendationBase.GetFromResourceRecommendationBase(recommendationList.Body);
+            responseRecommendationList = recommendationResourceUtil.GetAllRecommendationsFromClient(this.ResourceAdvisorClient);
 
             // Add the particular recommendation to the response of cmdlet
             responseRecommendation.Add(RecommendationHelper.RecommendationFilterByRecommendation(responseRecommendationList, recommendationId));
@@ -156,8 +154,8 @@ namespace Microsoft.Azure.Commands.Advisor.Cmdlets
 
                 case InputObjectParameterSet:
                     // Parse out the Subscription-ID, Recommendation-ID from the ResourceId parameter.
-                    resourceUri = RecommendationHelper.GetFullResourceUriFromResourceID(this.InputObject.Id);
-                    recommendationId = RecommendationHelper.GetRecommendationIdFromResourceID(this.InputObject.Id);
+                    resourceUri = RecommendationHelper.GetFullResourceUriFromResourceID(this.InputObject.ResourceId);
+                    recommendationId = RecommendationHelper.GetRecommendationIdFromResourceID(this.InputObject.ResourceId);
                     if (ShouldProcess(recommendationId, string.Format(Resources.EnableRecommendationWarningMessage, recommendationId)))
                     {
                         responseRecommendation.AddRange(this.SuppressionDelete(resourceUri, recommendationId));