+++ src/ResourceManager/Compute/Commands.Compute/ChangeLog.md
@@ -20,6 +20,7 @@
 -->
 ## Upcoming Release
 * Name is now optional in ID parameter set for Restart/Start/Stop/Remove/Set-AzVM and Save-AzVMImage
+* Updated the description of ID in help files
 
 ## Version 1.0.0
 * General availability of `Az.Compute` module